Added database version checks in upgradeDB method to apply only required changes to the database structure

+++ SWADroid/src/es/ugr/swad/swadroid/model/DataBaseHelper.java
@@ -43,6 +43,7 @@ import com.android.dataframework.DataFramework;
 import com.android.dataframework.Entity;
 
 import es.ugr.swad.swadroid.Global;
+//import es.ugr.swad.swadroid.Preferences;
 
 /**
  * @author Juan Miguel Boyero Corral <juanmi1982@gmail.com>
@@ -62,6 +63,10 @@ public class DataBaseHelper {
 	 * Application preferences
 	 */
 	//private Preferences prefs = new Preferences();
+	/**
+	 * Database name
+	 */
+	//private String DBName = "swadroid_db";
 	/**
 	 * Database passphrase
 	 */
@@ -88,10 +93,10 @@ public class DataBaseHelper {
 		/*if(DBKey.equals("")) {
 			DBKey = Global.randomString(DB_KEY_LENGTH);
 			prefs.setDBKey(DBKey);
-			mCtx.deleteDatabase("swadroid_db");*/
-			//SQLiteDatabase database = SQLiteDatabase.openOrCreateDatabase(mCtx.getDatabasePath("swadroid_db_crypt"), DBKey, null);
-			//database.close();
-		/*}
+			mCtx.deleteDatabase(DBName);
+			SQLiteDatabase database = SQLiteDatabase.openOrCreateDatabase(mCtx.getDatabasePath("swadroid_db_crypt"), DBKey, null);
+			database.close();
+		}
 		
 		Log.d("DataBaseHelper", "DBKey=" + DBKey);*/
 		
@@ -191,7 +196,9 @@ public class DataBaseHelper {
 					ent.getString("fullName"));
 		} else if(table.equals(Global.DB_TABLE_TEST_QUESTIONS_COURSE) ||
 				table.equals(Global.DB_TABLE_TEST_QUESTION_ANSWERS) ||
-				table.equals(Global.DB_TABLE_USERS_COURSES) || table.equals(Global.DB_TABLE_GROUPS_COURSES) || table.equals(Global.DB_TABLE_GROUPS_GROUPTYPES)) {
+				table.equals(Global.DB_TABLE_USERS_COURSES) ||
+				table.equals(Global.DB_TABLE_GROUPS_COURSES) ||
+				table.equals(Global.DB_TABLE_GROUPS_GROUPTYPES)) {
 
 			params = selectParamsPairTable(table);
 
@@ -1021,7 +1028,6 @@ public class DataBaseHelper {
 				Model model = newModels.get(i);
 				insertGroup((Group) model, courseCode[0]);
 			}
-			List<Entity> rows; 
 			for(int i = 0; i < modifiedModel.size(); ++i){
 				Model model = modifiedModel.get(i);
 				insertGroup((Group) model, courseCode[0]);
@@ -1759,7 +1765,10 @@ public class DataBaseHelper {
 	 * @throws IOException 
 	 * @throws XmlPullParserException 
 	 */
-	public void upgradeDB(Context context) throws XmlPullParserException, IOException {    	
+	public void upgradeDB(Context context) throws XmlPullParserException, IOException {
+		int dbVersion = db.getDB().getVersion();
+		boolean found = false;
+		int i = 0;
 		//cleanTables();    	
 		//initializeDB();
 
@@ -1781,39 +1790,39 @@ public class DataBaseHelper {
 		 * changes on courses table:
 		 * - old field name is erased
 		 * The rest of the changes are only new fields and they are added automatic by Dataframework. */
-		Cursor dbCursor = db.getDB().query(Global.DB_TABLE_COURSES, null, null, null, null, null, null);
-		String[] columnNames = dbCursor.getColumnNames();
-		boolean found = false;
-		int i = 0;
-		while(i < columnNames.length && !found){
-			if(columnNames[i].compareTo("name") == 0) found = true;
-			++i;			
-		}
-		if(found){
-			//without to keep data 
-			db.getDB().execSQL( "DROP TABLE " + Global.DB_TABLE_COURSES + ";");//+
-					db.getDB().execSQL("CREATE TABLE "+ Global.DB_TABLE_COURSES
-					+ " (_id integer primary key autoincrement, id long, userRole integer,shortName text, fullName text);");
-			//Keeping data (It will have columns without data):
-			/*
-			 * db.getDB().execSQL("CREATE TEMPORARY TABLE __"+ Global.DB_TABLE_COURSES
-			+ " (_id integer primary key autoincrement, id long, userRole integer,"
-            + " shortName text, fullName text); ");
-	db.getDB().execSQL(
-            "INSERT INTO __" + Global.DB_TABLE_COURSES + " SELECT _id, id, userRole, name, name  "
-           + " FROM "+ Global.DB_TABLE_COURSES + ";");
-	db.getDB().execSQL( "DROP TABLE " + Global.DB_TABLE_COURSES + ";");
-	db.getDB().execSQL("CREATE TABLE "+ Global.DB_TABLE_COURSES
-			+ " (_id integer primary key autoincrement, id long, userRole integer,"
-            + " shortName text, fullName text); ");
-	db.getDB().execSQL(
-            "INSERT INTO " + Global.DB_TABLE_COURSES + " SELECT _id, id, userRole, shortName, fullName  "
-           + " FROM __"+ Global.DB_TABLE_COURSES + ";");*/
-
-		}
-
-		dbCursor = db.getDB().query(Global.DB_TABLE_COURSES, null, null, null, null, null, null);
-		columnNames = dbCursor.getColumnNames();
+		if(dbVersion < 12) {
+			Cursor dbCursor = db.getDB().query(Global.DB_TABLE_COURSES, null, null, null, null, null, null);
+			String[] columnNames = dbCursor.getColumnNames();
+			while(i < columnNames.length && !found){
+				if(columnNames[i].compareTo("name") == 0) found = true;
+				++i;			
+			}
+			if(found){
+				//without to keep data 
+				db.getDB().execSQL( "DROP TABLE " + Global.DB_TABLE_COURSES + ";");//+
+						db.getDB().execSQL("CREATE TABLE "+ Global.DB_TABLE_COURSES
+						+ " (_id integer primary key autoincrement, id long, userRole integer,shortName text, fullName text);");
+				//Keeping data (It will have columns without data):
+				/*
+				 * db.getDB().execSQL("CREATE TEMPORARY TABLE __"+ Global.DB_TABLE_COURSES
+				+ " (_id integer primary key autoincrement, id long, userRole integer,"
+	            + " shortName text, fullName text); ");
+		db.getDB().execSQL(
+	            "INSERT INTO __" + Global.DB_TABLE_COURSES + " SELECT _id, id, userRole, name, name  "
+	           + " FROM "+ Global.DB_TABLE_COURSES + ";");
+		db.getDB().execSQL( "DROP TABLE " + Global.DB_TABLE_COURSES + ";");
+		db.getDB().execSQL("CREATE TABLE "+ Global.DB_TABLE_COURSES
+				+ " (_id integer primary key autoincrement, id long, userRole integer,"
+	            + " shortName text, fullName text); ");
+		db.getDB().execSQL(
+	            "INSERT INTO " + Global.DB_TABLE_COURSES + " SELECT _id, id, userRole, shortName, fullName  "
+	           + " FROM __"+ Global.DB_TABLE_COURSES + ";");*/
+	
+			}
+	
+			dbCursor = db.getDB().query(Global.DB_TABLE_COURSES, null, null, null, null, null, null);
+			columnNames = dbCursor.getColumnNames();
+			}
 
 		/* version 12 - 13
 		 * changes on groups table: 
@@ -1822,36 +1831,38 @@ public class DataBaseHelper {
 		 * - old field groupTypeName is erased
 		 * The rest of the changes are only new fields and they are added automatic by Dataframework. 
 		 * */
-		dbCursor = db.getDB().query(Global.DB_TABLE_GROUPS, null, null, null, null, null, null);
-		columnNames = dbCursor.getColumnNames();
-		found = false;
-		i = 0;
-		while(i < columnNames.length && !found){
-			if(columnNames[i].compareTo("groupCode") == 0) found = true;
-			++i;			
-		}
-		if(found){
-			//without to keep data 
+		if(dbVersion < 13) {
+			Cursor dbCursor = db.getDB().query(Global.DB_TABLE_COURSES, null, null, null, null, null, null);
+			String[] columnNames = dbCursor.getColumnNames();
+			dbCursor = db.getDB().query(Global.DB_TABLE_GROUPS, null, null, null, null, null, null);
+			columnNames = dbCursor.getColumnNames();
+			while(i < columnNames.length && !found){
+				if(columnNames[i].compareTo("groupCode") == 0) found = true;
+				++i;			
+			}
+			if(found){
+				//without to keep data 
+				db.getDB().execSQL( "DROP TABLE " + Global.DB_TABLE_GROUPS + ";");
+				db.getDB().execSQL("CREATE TABLE " + Global.DB_TABLE_GROUPS+ " (_id integer primary key autoincrement, id long, groupName text, maxStudents integer,"
+		                + " students integer, open integer, fileZones integer, member integer); ");
+			/*db.getDB().execSQL(
+					"CREATE TEMPORARY TABLE __"+ Global.DB_TABLE_GROUPS
+					+ " (_id integer primary key autoincrement, id long, groupName text, maxStudents integer,"
+	                + " students integer, open integer, fileZones integer, member integer); ");
+			db.getDB().execSQL(
+	                 "INSERT INTO __" + Global.DB_TABLE_GROUPS + " SELECT _id, groupCode, groupName, maxStudents,  "
+	                + "students, open, fileZones, member FROM "+ Global.DB_TABLE_GROUPS + ";");
 			db.getDB().execSQL( "DROP TABLE " + Global.DB_TABLE_GROUPS + ";");
 			db.getDB().execSQL("CREATE TABLE " + Global.DB_TABLE_GROUPS+ " (_id integer primary key autoincrement, id long, groupName text, maxStudents integer,"
 	                + " students integer, open integer, fileZones integer, member integer); ");
-		/*db.getDB().execSQL(
-				"CREATE TEMPORARY TABLE __"+ Global.DB_TABLE_GROUPS
-				+ " (_id integer primary key autoincrement, id long, groupName text, maxStudents integer,"
-                + " students integer, open integer, fileZones integer, member integer); ");
-		db.getDB().execSQL(
-                 "INSERT INTO __" + Global.DB_TABLE_GROUPS + " SELECT _id, groupCode, groupName, maxStudents,  "
-                + "students, open, fileZones, member FROM "+ Global.DB_TABLE_GROUPS + ";");
-		db.getDB().execSQL( "DROP TABLE " + Global.DB_TABLE_GROUPS + ";");
-		db.getDB().execSQL("CREATE TABLE " + Global.DB_TABLE_GROUPS+ " (_id integer primary key autoincrement, id long, groupName text, maxStudents integer,"
-                + " students integer, open integer, fileZones integer, member integer); ");
-        db.getDB().execSQL("INSERT INTO " + Global.DB_TABLE_GROUPS + " SELECT _id, id, groupName, maxStudents,  "
-                + "students, open, fileZones, member FROM __"+ Global.DB_TABLE_GROUPS + ";");
-        db.getDB().execSQL( "DROP TABLE __" + Global.DB_TABLE_GROUPS + ";");*/
-
+	        db.getDB().execSQL("INSERT INTO " + Global.DB_TABLE_GROUPS + " SELECT _id, id, groupName, maxStudents,  "
+	                + "students, open, fileZones, member FROM __"+ Global.DB_TABLE_GROUPS + ";");
+	        db.getDB().execSQL( "DROP TABLE __" + Global.DB_TABLE_GROUPS + ";");*/
+	
+			}
+			dbCursor = db.getDB().query(Global.DB_TABLE_GROUPS, null, null, null, null, null, null);
+			columnNames = dbCursor.getColumnNames();
 		}
-		dbCursor = db.getDB().query(Global.DB_TABLE_GROUPS, null, null, null, null, null, null);
-		columnNames = dbCursor.getColumnNames();
 
 		/*db.getDB().execSQL("CREATE TEMPORARY TABLE __"
                 + Global.DB_TABLE_NOTIFICATIONS
